// src/app/models/admin/capphephoatdongkhoangsan/cpdaugiakhaithac/cpdaugiakhuvuc.model.ts
export class InputCpDauGiaKhuVucModel {
  iddaugiakhuvuc: string;
  tenkhuvuc: string;
  dientich: number;
  donvidientich: string;
  idcapphepdaugia: string;
  loaikhuvuc: number;
  loaicapphep: string;
  hequychieu: string;
  lstToado: any;

}

export class OutputCpDauGiaKhuVucModel {
  iddaugiakhuvuc: string;
  tenkhuvuc: string;
  dientich: number;
  donvidientich: string;
  idcapphepdaugia: string;
  loaikhuvuc: number;
  loaicapphep: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  siteid: string;
  lstToado: any;
}